Core classes and some file IO
Added the Character, Inventory, and Item classes.  Basic stats for characters.  This can be done by constructor or from a file in the characters/ folder (see liora_virelle.json).  I don't like that I have to use dot-notation for the imports in the res folder just now, but it is what it is.